# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Project API",
      "description": "API for managing construction projects",
      "version": "1.0.0"
    },
    "host": "localhost:3001",
    "basePath": "/api",
    "paths": {
      "/projects": {
        "get": {
          "summary": "Get all projects",
          "responses": {
            "200": {
              "description": "List of projects",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Project"
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new project",
          "parameters": [
            {
              "in": "body",
              "name": "project",
              "description": "Project object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Project"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Project created successfully",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            }
          }
        }
      },
      "/projects/{id}": {
        "get": {
          "summary": "Get project by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Project details",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            }
          }
        },
        "put": {
          "summary": "Update project by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "project",
              "description": "Updated project object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Project"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Project updated successfully",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete project by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Project deleted successfully"
            }
          }
        }
      }
    },
    "definitions": {
      "Project": {
        "type": "object",
        "properties": {
          "title": {
            "type": "object",
            "properties": {
              "en": { "type": "string" },
              "ar": { "type": "string" }
            }
          },
          "slogan": {
            "type": "object",
            "properties": {
              "en": { "type": "string" },
              "ar": { "type": "string" }
            }
          },
          "description": {
            "type": "object",
            "properties": {
              "en": { "type": "string" },
              "ar": { "type": "string" }
            }
          },
          "location": {
            "type": "object",
            "properties": {
              "en": { "type": "string" },
              "ar": { "type": "string" }
            }
          },
          "thumbnail": { "type": "string" },
          "images": {
            "type": "array",
            "items": { "type": "string" }
          },
          "building_area": { "type": "number" },
          "land_area": { "type": "number" },
          "units": { "type": "number" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "label": {
            "type": "object",
            "properties": {
              "en": { "type": "string" },
              "ar": { "type": "string" }
            }
          }
        }
      }
    }
  }
  